+++ CHANGES.md
@@ -1,5 +1,6 @@
 **Changed:**
 
+- Development instructions in project templates were moved from `README.md` to a `CONTRIBUTING.md`, and are now documented using `npm` and `npm run` commands instead of global `nwb` commands.
 - All commands are now run in the current working directory - you no longer need to `require.resolve()` full paths to extra Babel plugins configured in `nwb.config.js`, just use their names as Babel will now be able to import them.
 
 # 0.7.1 / 2016-01-10